+++ src/main/java/gyro/aws/ec2/LaunchTemplateResource.java
@@ -34,6 +34,7 @@ import gyro.core.resource.Id;
 import gyro.core.resource.Output;
 import gyro.core.resource.Updatable;
 import gyro.core.scope.State;
+import gyro.core.validation.DependsOn;
 import gyro.core.validation.Required;
 import org.apache.commons.codec.binary.Base64;
 import software.amazon.awssdk.services.ec2.Ec2Client;
@@ -43,11 +44,9 @@ import software.amazon.awssdk.services.ec2.model.DescribeLaunchTemplatesResponse
 import software.amazon.awssdk.services.ec2.model.Ec2Exception;
 import software.amazon.awssdk.services.ec2.model.InstanceType;
 import software.amazon.awssdk.services.ec2.model.LaunchTemplate;
-import software.amazon.awssdk.services.ec2.model.LaunchTemplateCapacityReservationSpecificationRequest;
 import software.amazon.awssdk.services.ec2.model.LaunchTemplateIamInstanceProfileSpecificationRequest;
 import software.amazon.awssdk.services.ec2.model.LaunchTemplateInstanceNetworkInterfaceSpecificationRequest;
 import software.amazon.awssdk.services.ec2.model.RequestLaunchTemplateData;
-import software.amazon.awssdk.utils.builder.SdkBuilder;
 
 /**
  * Creates a Launch Template from config or an existing Instance Id.
@@ -94,22 +93,31 @@ public class LaunchTemplateResource extends Ec2TaggableResource<LaunchTemplate> 
 
     private String name;
     private AmiResource ami;
-    private Integer coreCount;
-    private Integer threadPerCore;
     private Boolean ebsOptimized;
     private Boolean configureHibernateOption;
     private String shutdownBehavior;
     private String instanceType;
     private String keyName;
-    private Boolean enableMonitoring;
     private List<SecurityGroupResource> securityGroups;
     private Boolean disableApiTermination;
     private String userData;
     private List<BlockDeviceMapping> blockDeviceMapping;
-    private String capacityReservation;
+    private LaunchTemplateCapacityReservation capacityReservation;
     private InstanceProfileResource instanceProfile;
     private Set<NetworkInterfaceResource> networkInterfaces;
     private LaunchTemplateMetadataOptions metadataOptions;
+    private LaunchTemplateCreditSpecification creditSpecification;
+    private List<LaunchTemplateElasticGpuSpecification> elasticGpuSpecification;
+    private List<LaunchTemplateElasticInferenceAccelerator> inferenceAccelerator;
+    private LaunchTemplateEnclaveOptions enclaveOptions;
+    private LaunchTemplateHibernationOptions hibernationOptions;
+    private LaunchTemplateInstanceMarketOptions marketOptions;
+    private String kernelId;
+    private String ramDiskId;
+    private LaunchTemplatePlacement placement;
+    private List<LaunchTemplateTagSpecification> tagSpecification;
+    private LaunchTemplateCpuOptions cpuOptions;
+    private LaunchTemplateMonitoring monitoring;
 
     // Read-only
     private String id;
@@ -137,38 +145,6 @@ public class LaunchTemplateResource extends Ec2TaggableResource<LaunchTemplate> 
 
     public void setAmi(AmiResource ami) {
         this.ami = ami;
-    }
-
-    /**
-     * The number of cores with which to launch instances. Defaults to 0 which sets its to the instance type defaults. See `Optimizing CPU Options <https://docs.aws.amazon.com/AWSEC2/latest/UserGuide/instance-optimize-cpu.html/>`_.
-     */
-    @Updatable
-    public Integer getCoreCount() {
-        if (coreCount == null) {
-            coreCount = 0;
-        }
-
-        return coreCount;
-    }
-
-    public void setCoreCount(Integer coreCount) {
-        this.coreCount = coreCount;
-    }
-
-    /**
-     * The number of threads per cores with which to launch instances. Defaults to 0 which sets its to the instance type defaults. See `Optimizing CPU Options <https://docs.aws.amazon.com/AWSEC2/latest/UserGuide/instance-optimize-cpu.html/>`_.
-     */
-    @Updatable
-    public Integer getThreadPerCore() {
-        if (threadPerCore == null) {
-            threadPerCore = 0;
-        }
-
-        return threadPerCore;
-    }
-
-    public void setThreadPerCore(Integer threadPerCore) {
-        this.threadPerCore = threadPerCore;
     }
 
     /**
@@ -238,21 +214,6 @@ public class LaunchTemplateResource extends Ec2TaggableResource<LaunchTemplate> 
 
     public void setKeyName(String keyName) {
         this.keyName = keyName;
-    }
-
-    /**
-     * When set to ``true``, monitoring for your instance is enabled. See `Monitoring Your Instances <https://docs.aws.amazon.com/AWSEC2/latest/UserGuide/using-cloudwatch.html/>`_.
-     */
-    @Updatable
-    public Boolean getEnableMonitoring() {
-        if (enableMonitoring == null) {
-            enableMonitoring = false;
-        }
-        return enableMonitoring;
-    }
-
-    public void setEnableMonitoring(Boolean enableMonitoring) {
-        this.enableMonitoring = enableMonitoring;
     }
 
     /**
@@ -327,15 +288,11 @@ public class LaunchTemplateResource extends Ec2TaggableResource<LaunchTemplate> 
      * The capacity reservation for the instances being launched using this template.
      */
     @Updatable
-    public String getCapacityReservation() {
-        if (capacityReservation == null) {
-            capacityReservation = "none";
-        }
-
+    public LaunchTemplateCapacityReservation getCapacityReservation() {
         return capacityReservation;
     }
 
-    public void setCapacityReservation(String capacityReservation) {
+    public void setCapacityReservation(LaunchTemplateCapacityReservation capacityReservation) {
         this.capacityReservation = capacityReservation;
     }
 
@@ -379,6 +336,165 @@ public class LaunchTemplateResource extends Ec2TaggableResource<LaunchTemplate> 
 
     public void setMetadataOptions(LaunchTemplateMetadataOptions metadataOptions) {
         this.metadataOptions = metadataOptions;
+    }
+
+    /**
+     * The credit specifications for the instance.
+     */
+    @Updatable
+    public LaunchTemplateCreditSpecification getCreditSpecification() {
+        return creditSpecification;
+    }
+
+    public void setCreditSpecification(LaunchTemplateCreditSpecification creditSpecification) {
+        this.creditSpecification = creditSpecification;
+    }
+
+    /**
+     * The elastic GPU to associate with the instance.
+     */
+    @Updatable
+    public List<LaunchTemplateElasticGpuSpecification> getElasticGpuSpecification() {
+        if (elasticGpuSpecification == null) {
+            elasticGpuSpecification = new ArrayList<>();
+        }
+
+        return elasticGpuSpecification;
+    }
+
+    public void setElasticGpuSpecification(List<LaunchTemplateElasticGpuSpecification> elasticGpuSpecification) {
+        this.elasticGpuSpecification = elasticGpuSpecification;
+    }
+
+    /**
+     * The elastic inference accelerator for the instance.
+     */
+    @Updatable
+    public List<LaunchTemplateElasticInferenceAccelerator> getInferenceAccelerator() {
+        if (inferenceAccelerator == null) {
+            inferenceAccelerator = new ArrayList<>();
+        }
+
+        return inferenceAccelerator;
+    }
+
+    public void setInferenceAccelerator(List<LaunchTemplateElasticInferenceAccelerator> inferenceAccelerator) {
+        this.inferenceAccelerator = inferenceAccelerator;
+    }
+
+    /**
+     * The enclave options for the instance.
+     */
+    @Updatable
+    public LaunchTemplateEnclaveOptions getEnclaveOptions() {
+        return enclaveOptions;
+    }
+
+    public void setEnclaveOptions(LaunchTemplateEnclaveOptions enclaveOptions) {
+        this.enclaveOptions = enclaveOptions;
+    }
+
+    /**
+     * The hibernation options for the instance.
+     */
+    @Updatable
+    public LaunchTemplateHibernationOptions getHibernationOptions() {
+        return hibernationOptions;
+    }
+
+    public void setHibernationOptions(LaunchTemplateHibernationOptions hibernationOptions) {
+        this.hibernationOptions = hibernationOptions;
+    }
+
+    /**
+     * The spot market options for the instance.
+     */
+    @Updatable
+    public LaunchTemplateInstanceMarketOptions getMarketOptions() {
+        return marketOptions;
+    }
+
+    public void setMarketOptions(LaunchTemplateInstanceMarketOptions marketOptions) {
+        this.marketOptions = marketOptions;
+    }
+
+    /**
+     * The tags to apply to the resources during launch.
+     *
+     * @subresource gyro.aws.ec2.LaunchTemplateTagSpecification
+     */
+    @Updatable
+    public List<LaunchTemplateTagSpecification> getTagSpecification() {
+        if (tagSpecification == null) {
+            tagSpecification = new ArrayList<>();
+        }
+
+        return tagSpecification;
+    }
+
+    public void setTagSpecification(List<LaunchTemplateTagSpecification> tagSpecification) {
+        this.tagSpecification = tagSpecification;
+    }
+
+    /**
+     * The ID of the kernel.
+     */
+    @Updatable
+    public String getKernelId() {
+        return kernelId;
+    }
+
+    public void setKernelId(String kernelId) {
+        this.kernelId = kernelId;
+    }
+
+    /**
+     * The ID of the RAM disk.
+     */
+    @Updatable
+    @DependsOn("kernel")
+    public String getRamDiskId() {
+        return ramDiskId;
+    }
+
+    public void setRamDiskId(String ramDiskId) {
+        this.ramDiskId = ramDiskId;
+    }
+
+    /**
+     * The placement for the instance;
+     */
+    @Updatable
+    public LaunchTemplatePlacement getPlacement() {
+        return placement;
+    }
+
+    public void setPlacement(LaunchTemplatePlacement placement) {
+        this.placement = placement;
+    }
+
+    /**
+     * The cpu options for the instance.
+     */
+    @Updatable
+    public LaunchTemplateCpuOptions getCpuOptions() {
+        return cpuOptions;
+    }
+
+    public void setCpuOptions(LaunchTemplateCpuOptions cpuOptions) {
+        this.cpuOptions = cpuOptions;
+    }
+
+    /**
+     * The monitoring for the instance.
+     */
+    @Updatable
+    public LaunchTemplateMonitoring getMonitoring() {
+        return monitoring;
+    }
+
+    public void setMonitoring(LaunchTemplateMonitoring monitoring) {
+        this.monitoring = monitoring;
     }
 
     /**
@@ -481,13 +597,6 @@ public class LaunchTemplateResource extends Ec2TaggableResource<LaunchTemplate> 
         if (!getSecurityGroups().isEmpty() && !getNetworkInterfaces().isEmpty()) {
             throw new GyroException("Either security group or network interface is to be provided, not both.");
         }
-
-        if (!getCapacityReservation().equalsIgnoreCase("none")
-            && !getCapacityReservation().equalsIgnoreCase("open")
-            && !getCapacityReservation().startsWith("cr-")) {
-            throw new GyroException("The value - (" + getCapacityReservation() + ") is invalid for parameter 'capacity-reservation'. "
-                + "Valid values [ 'open', 'none', capacity reservation id like cr-% ]");
-        }
     }
 
     private LaunchTemplate getLaunchTemplate(Ec2Client client) {
@@ -514,18 +623,6 @@ public class LaunchTemplateResource extends Ec2TaggableResource<LaunchTemplate> 
         return launchTemplate;
     }
 
-    private LaunchTemplateCapacityReservationSpecificationRequest getCapacityReservationSpecification() {
-        if (("none").equals(getCapacityReservation()) || ("open").equals(getCapacityReservation())) {
-            return LaunchTemplateCapacityReservationSpecificationRequest.builder()
-                .capacityReservationPreference(getCapacityReservation().toLowerCase())
-                .build();
-        } else {
-            return LaunchTemplateCapacityReservationSpecificationRequest.builder()
-                .capacityReservationTarget(r -> r.capacityReservationId(getCapacityReservation()))
-                .build();
-        }
-    }
-
     private LaunchTemplateIamInstanceProfileSpecificationRequest getLaunchTemplateInstanceProfile() {
         if (getInstanceProfile() == null) {
             return null;
@@ -544,10 +641,7 @@ public class LaunchTemplateResource extends Ec2TaggableResource<LaunchTemplate> 
     }
 
     private RequestLaunchTemplateData requestLaunchTemplateData() {
-        return RequestLaunchTemplateData.builder()
-            .cpuOptions(getCoreCount() > 0
-            ? o -> o.threadsPerCore(getThreadPerCore())
-            .coreCount(getCoreCount()).build() : SdkBuilder::build)
+        RequestLaunchTemplateData.Builder builder = RequestLaunchTemplateData.builder()
             .disableApiTermination(getDisableApiTermination())
             .ebsOptimized(getEbsOptimized())
             .hibernationOptions(o -> o.configured(getConfigureHibernateOption()))
@@ -555,7 +649,6 @@ public class LaunchTemplateResource extends Ec2TaggableResource<LaunchTemplate> 
             .instanceType(getInstanceType())
             .instanceInitiatedShutdownBehavior(getShutdownBehavior())
             .keyName(getKeyName())
-            .monitoring(o -> o.enabled(getEnableMonitoring()))
             .securityGroupIds(!getSecurityGroups().isEmpty() ? getSecurityGroups().stream()
                 .map(SecurityGroupResource::getId)
                 .collect(Collectors.toList()) : null)
@@ -565,11 +658,61 @@ public class LaunchTemplateResource extends Ec2TaggableResource<LaunchTemplate> 
                     .stream().map(BlockDeviceMapping::getLaunchTemplateBlockDeviceMapping)
                     .collect(Collectors.toList()) : null
             )
-            .capacityReservationSpecification(getCapacityReservationSpecification())
+            .capacityReservationSpecification(getCapacityReservation() == null ? null
+                : getCapacityReservation().toLaunchTemplateCapacityReservationSpecificationRequest())
             .iamInstanceProfile(getLaunchTemplateInstanceProfile())
             .networkInterfaces(!getNetworkInterfaces().isEmpty()
                 ? toNetworkInterfaceSpecificationRequest() : null)
-            .metadataOptions(getMetadataOptions() == null ? null : getMetadataOptions().toMetadataOptions()
-            ).build();
+            .kernelId(getKernelId())
+            .ramDiskId(getRamDiskId())
+            .metadataOptions(getMetadataOptions() == null ? null : getMetadataOptions().toMetadataOptions());
+
+        if (getCreditSpecification() != null) {
+            builder.creditSpecification(getCreditSpecification().toCreditSpecification());
+        }
+
+        if (!getElasticGpuSpecification().isEmpty()) {
+            builder.elasticGpuSpecifications(getElasticGpuSpecification().stream()
+                .map(LaunchTemplateElasticGpuSpecification::toElasticGpuSpecification)
+                .collect(Collectors.toList()));
+        }
+
+        if (!getInferenceAccelerator().isEmpty()) {
+            builder.elasticInferenceAccelerators(getInferenceAccelerator().stream()
+                .map(LaunchTemplateElasticInferenceAccelerator::toLaunchTemplateElasticInferenceAccelerator)
+                .collect(Collectors.toList()));
+        }
+
+        if (getEnclaveOptions() != null) {
+            builder.enclaveOptions(getEnclaveOptions().toLaunchTemplateEnclaveOptionsRequest());
+        }
+
+        if (getHibernationOptions() != null) {
+            builder.hibernationOptions(getHibernationOptions().toLaunchTemplateHibernationOptionsRequest());
+        }
+
+        if (getMarketOptions() != null) {
+            builder.instanceMarketOptions(getMarketOptions().toLaunchTemplateInstanceMarketOptionsRequest());
+        }
+
+        if (getPlacement() != null) {
+            builder.placement(getPlacement().toLaunchTemplatePlacementRequest());
+        }
+
+        if (!getTagSpecification().isEmpty()) {
+            builder.tagSpecifications(getTagSpecification().stream()
+                .map(LaunchTemplateTagSpecification::toLaunchTemplateTagSpecificationRequest)
+                .collect(Collectors.toList()));
+        }
+
+        if (getCpuOptions() != null) {
+            builder.cpuOptions(getCpuOptions().toLaunchTemplateCpuOptionsRequest());
+        }
+
+        if (getMonitoring() != null) {
+            builder.monitoring(getMonitoring().toLaunchTemplatesMonitoringRequest()).build();
+        }
+
+        return builder.build();
     }
 }