+++ src/main/java/gyro/aws/eks/EksAddonResource.java
@@ -25,7 +25,6 @@ import gyro.aws.AwsResource;
 import gyro.aws.Copyable;
 import gyro.aws.iam.RoleResource;
 import gyro.core.GyroUI;
-import gyro.core.Type;
 import gyro.core.Wait;
 import gyro.core.resource.Id;
 import gyro.core.resource.Output;
@@ -49,21 +48,23 @@ import software.amazon.awssdk.services.eks.model.UntagResourceRequest;
  *
  * .. code-block:: gyro
  *
- *     aws::eks-addon example-addon
- *         addon-name: vpc-cni
- *         cluster: $(aws::eks-cluster ex)
+ *     aws::eks-cluster example-addon
+ *         .
+ *         .
+ *         .
+ *         addon
+ *              addon-name: vpc-cni
  *
- *         tags : {
- *             "example-tag-key": "example-key-value"
- *         }
+ *              tags : {
+ *                  Name: "example-key-value"
+ *              }
+ *         end
  *     end
  */
-@Type("eks-addon")
 public class EksAddonResource extends AwsResource implements Copyable<Addon> {
 
     private String addonName;
     private String addonVersion;
-    private EksClusterResource cluster;
     private RoleResource serviceAccountRole;
     private Map<String, String> tags;
 
@@ -74,7 +75,6 @@ public class EksAddonResource extends AwsResource implements Copyable<Addon> {
      * The name of the add-on.
      */
     @Required
-    @Id
     public String getAddonName() {
         return addonName;
     }
@@ -93,18 +93,6 @@ public class EksAddonResource extends AwsResource implements Copyable<Addon> {
 
     public void setAddonVersion(String addonVersion) {
         this.addonVersion = addonVersion;
-    }
-
-    /**
-     * The name of the cluster to create the add-on for.
-     */
-    @Required
-    public EksClusterResource getCluster() {
-        return cluster;
-    }
-
-    public void setCluster(EksClusterResource cluster) {
-        this.cluster = cluster;
     }
 
     /**
@@ -148,10 +136,14 @@ public class EksAddonResource extends AwsResource implements Copyable<Addon> {
     }
 
     @Override
+    public String primaryKey() {
+        return getAddonName();
+    }
+
+    @Override
     public void copyFrom(Addon model) {
         setAddonName(model.addonName());
         setAddonVersion(model.addonVersion());
-        setCluster(findById(EksClusterResource.class, model.clusterName()));
         setServiceAccountRole(findById(RoleResource.class, model.serviceAccountRoleArn()));
         setArn(model.addonArn());
         setTags(model.tags());
@@ -159,16 +151,6 @@ public class EksAddonResource extends AwsResource implements Copyable<Addon> {
 
     @Override
     public boolean refresh() {
-        EksClient client = createClient(EksClient.class);
-
-        Addon addon = getAddon(client);
-
-        if (addon == null || addon.status().equals(AddonStatus.DELETING)) {
-            return false;
-        }
-
-        copyFrom(addon);
-
         return true;
     }
 
@@ -176,15 +158,17 @@ public class EksAddonResource extends AwsResource implements Copyable<Addon> {
     public void create(GyroUI ui, State state) throws Exception {
         EksClient client = createClient(EksClient.class);
 
+        EksClusterResource parent = (EksClusterResource) parent();
+
         CreateAddonResponse response = client.createAddon(r -> r.addonName(getAddonName())
             .addonVersion(getAddonVersion())
-            .clusterName(getCluster().getName())
+            .clusterName(parent.getName())
             .serviceAccountRoleArn(getServiceAccountRole() == null ? null : getServiceAccountRole().getArn())
             .tags(getTags()));
 
         setArn(response.addon().addonArn());
 
-        waitForActiveStatus(client);
+        waitForActiveStatus(client, parent.getName(), getAddonName());
 
         copyFrom(response.addon());
     }
@@ -192,6 +176,8 @@ public class EksAddonResource extends AwsResource implements Copyable<Addon> {
     @Override
     public void update(GyroUI ui, State state, Resource current, Set<String> changedFieldNames) throws Exception {
         EksClient client = createClient(EksClient.class);
+
+        EksClusterResource parent = (EksClusterResource) parent();
 
         if (changedFieldNames.contains("tags")) {
             EksAddonResource currentResource = (EksAddonResource) current;
@@ -209,10 +195,10 @@ public class EksAddonResource extends AwsResource implements Copyable<Addon> {
         if (changedFieldNames.contains("addon-version") || changedFieldNames.contains("service-account-role")) {
             client.updateAddon(r -> r.addonName(getAddonName())
                 .addonVersion(getAddonVersion())
-                .clusterName(getCluster().getName())
+                .clusterName(parent.getName())
                 .serviceAccountRoleArn(getServiceAccountRole() == null ? null : getServiceAccountRole().getArn()));
 
-            waitForActiveStatus(client);
+            waitForActiveStatus(client, parent.getName(), getAddonName());
         }
     }
 
@@ -220,19 +206,21 @@ public class EksAddonResource extends AwsResource implements Copyable<Addon> {
     public void delete(GyroUI ui, State state) throws Exception {
         EksClient client = createClient(EksClient.class);
 
-        client.deleteAddon(r -> r.addonName(getAddonName()).clusterName(getCluster().getName()));
+        EksClusterResource parent = (EksClusterResource) parent();
+
+        client.deleteAddon(r -> r.addonName(getAddonName()).clusterName(parent.getName()));
 
         Wait.atMost(1, TimeUnit.MINUTES)
             .checkEvery(10, TimeUnit.SECONDS)
             .prompt(false)
-            .until(() -> getAddon(client) == null);
+            .until(() -> getAddon(client, parent.getName(), getAddonName()) == null);
     }
 
-    private Addon getAddon(EksClient client) {
+    protected static Addon getAddon(EksClient client, String clusterName, String name) {
         Addon addon = null;
 
         try {
-            addon = client.describeAddon(r -> r.addonName(getAddonName()).clusterName(getCluster().getName())).addon();
+            addon = client.describeAddon(r -> r.addonName(name).clusterName(clusterName)).addon();
 
         } catch (ResourceNotFoundException ex) {
             // ignore
@@ -241,12 +229,12 @@ public class EksAddonResource extends AwsResource implements Copyable<Addon> {
         return addon;
     }
 
-    private void waitForActiveStatus(EksClient client) {
+    private void waitForActiveStatus(EksClient client, String clusterName, String name) {
         Wait.atMost(2, TimeUnit.MINUTES)
             .checkEvery(10, TimeUnit.SECONDS)
             .prompt(false)
             .until(() -> {
-                Addon addon = getAddon(client);
+                Addon addon = getAddon(client, clusterName, name);
                 return addon != null && addon.status().equals(AddonStatus.ACTIVE);
             });
     }