+++ src/main/java/gyro/aws/eks/EksClusterResource.java
@@ -37,6 +37,7 @@ import gyro.core.resource.Updatable;
 import gyro.core.scope.State;
 import gyro.core.validation.Required;
 import software.amazon.awssdk.services.eks.EksClient;
+import software.amazon.awssdk.services.eks.model.Addon;
 import software.amazon.awssdk.services.eks.model.Cluster;
 import software.amazon.awssdk.services.eks.model.ClusterStatus;
 import software.amazon.awssdk.services.eks.model.CreateClusterRequest;
@@ -44,9 +45,14 @@ import software.amazon.awssdk.services.eks.model.CreateClusterResponse;
 import software.amazon.awssdk.services.eks.model.DeleteClusterRequest;
 import software.amazon.awssdk.services.eks.model.DescribeClusterRequest;
 import software.amazon.awssdk.services.eks.model.EksException;
+import software.amazon.awssdk.services.eks.model.ListAddonsResponse;
+import software.amazon.awssdk.services.eks.model.IdentityProviderConfig;
+import software.amazon.awssdk.services.eks.model.IdentityProviderConfigResponse;
+import software.amazon.awssdk.services.eks.model.ListIdentityProviderConfigsResponse;
 import software.amazon.awssdk.services.eks.model.LogSetup;
 import software.amazon.awssdk.services.eks.model.LogType;
 import software.amazon.awssdk.services.eks.model.Logging;
+import software.amazon.awssdk.services.eks.model.NotFoundException;
 import software.amazon.awssdk.services.eks.model.TagResourceRequest;
 import software.amazon.awssdk.services.eks.model.UntagResourceRequest;
 import software.amazon.awssdk.services.eks.model.UpdateClusterConfigRequest;
@@ -110,6 +116,8 @@ public class EksClusterResource extends AwsResource implements Copyable<Cluster>
     private EksLogging logging;
     private List<EksEncryptionConfig> encryptionConfig;
     private Map<String, String> tags;
+    private List<EksAddonResource> addon;
+    private EksAuthentication authentication;
 
     // Read-only
     private String arn;
@@ -202,6 +210,24 @@ public class EksClusterResource extends AwsResource implements Copyable<Cluster>
     }
 
     /**
+     * The addon configuration for the cluster.
+     *
+     * @subresource gyro.aws.eks.EksAddonResource
+     */
+    @Updatable
+    public List<EksAddonResource> getAddon() {
+        if (addon == null) {
+            addon = new ArrayList<>();
+        }
+
+        return addon;
+    }
+
+    public void setAddon(List<EksAddonResource> addon) {
+        this.addon = addon;
+    }
+
+    /**
      * The tags to attach to the cluster.
      */
     @Updatable
@@ -215,6 +241,20 @@ public class EksClusterResource extends AwsResource implements Copyable<Cluster>
 
     public void setTags(Map<String, String> tags) {
         this.tags = tags;
+    }
+
+    /**
+     * The authentication config for the cluster.
+     *
+     * @subresource gyro.aws.eks.EksAuthentication
+     */
+    @Updatable
+    public EksAuthentication getAuthentication() {
+        return authentication;
+    }
+
+    public void setAuthentication(EksAuthentication authentication) {
+        this.authentication = authentication;
     }
 
     /**
@@ -287,6 +327,51 @@ public class EksClusterResource extends AwsResource implements Copyable<Cluster>
 
         if (model.certificateAuthority() != null) {
             setCertificateAuthorityData(model.certificateAuthority().data());
+        }
+
+        EksClient client = createClient(EksClient.class);
+
+        // load addon
+        setAddon(null);
+        try {
+            ListAddonsResponse response = client.listAddons(r -> r.clusterName(getName()));
+
+            if (response.hasAddons()) {
+                response.addons().forEach(a -> {
+                    Addon addon = EksAddonResource.getAddon(client, getName(), a);
+                    if (addon != null) {
+                        EksAddonResource addonResource = newSubresource(EksAddonResource.class);
+                        addonResource.copyFrom(addon);
+                        getAddon().add(addonResource);
+                    }
+                });
+            }
+        } catch (NotFoundException ex) {
+            // Ignore
+        }
+
+        // load eks authentication
+        setAuthentication(null);
+        try {
+            ListIdentityProviderConfigsResponse response = client.listIdentityProviderConfigs(r -> r
+                .clusterName(getName()));
+            if (response.hasIdentityProviderConfigs() && !response.identityProviderConfigs().isEmpty()) {
+                IdentityProviderConfig providerConfig = response.identityProviderConfigs().get(0);
+
+                IdentityProviderConfigResponse auth = EksAuthentication.getIdentityProviderConfigResponse(
+                    client,
+                    getName(),
+                    providerConfig.name(),
+                    providerConfig.type());
+
+                if (auth != null) {
+                    EksAuthentication authentication = newSubresource(EksAuthentication.class);
+                    authentication.copyFrom(auth);
+                    setAuthentication(authentication);
+                }
+            }
+        } catch (NotFoundException ex) {
+            // Ignore
         }
     }